Store calibration picks in Cartesian coordinates
Previously, the calibration picks were being stored in angular coordinates.
This was a problem, however, because if the instrument/detector parameters
were modified, the picks would not be updated to reflect that change.

This commit changes the calibration picks to be stored in Cartesian
coordinates instead (similar to how they are stored in the HDF5 file).
And to alleviate the transition, a `calibration_picks_polar` property was
added that automatically converts between the polar/cartesian coordinates.

Fixes: #1310

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>