Add extra sample points for masks
For masks with few points (such as the rectangle mask, which only had
four corners for the points, and hand drawn masks with few points),
the masks would not be converted between the polar/raw views very well.

For example, for the rectangle mask, the four corners would be converted
between the views correctly. But the edges in the converted form would be
straight, when they would often need to be curved.

We need to add extra sample points for the masks so that conversions from
straight edges to curved edges are performed correctly. This commit does
that.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>