Fix convert_panel_buffer_to_2d_array()
The 2-value panel buffer represents the `x, y` directions in mm,
not in pixels.

So we need to convert these values to pixels first, and then we also needed
to fix an issue where the logic wouldn't work if the borders were 0. This
would result in the whole array being `False`. That is now fixed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>