Add memoization for faster polar view generation
This adds memoization for projecting onto the detector plane and
for applying SNIP. This makes it much faster to switch back and forth
between different views (without changing the settings). And this also
makes the stereo generation from the polar view much faster, since most
of the slow steps for generating the polar view are memoized.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>