Make a deep copy of materials in indexing config
HEXRD sometimes modifies these materials. This is why we were originally
setting tth width on the config, so that our material would not be modified.

However, HEXRD also sometimes modifies other properties. So make a deep copy
of the materials so that they do not get uninentionally modified in HEXRDGUI.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>