Skip converting fast powder picks to polar
We are now storing the calibration picks as Cartesian coordinates -
not polar coordinates. As such, we do not need to convert to polar
coordinates anymore, but can set the Cartesian picks directly.

This fixes an issue where, if you were looking at the Cartesian view,
the auto picks would fail to convert to polar coordinates (due to the
fact that the Cartesian view has a fake instrument it uses).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>