Add a utility method for opening help URLs
This adds a utility method that takes a button box for the first
argument and a URL for the second argument.

This adds a connection so that if the "Help" button is pressed,
then the specified URL will be opened.

If the button box doesn't have a "Help" button, one will be added.

This also adds an example of it in action for the simple image series,
where we link the simple image series page on the docs.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>