Add ability to compute polar corr field directly
This allows us to skip the step of creating the tth displacement fields
for the detectors themselves, and instead directly create the tth displacement
field for the polar view, if that is our goal.

This provides a huge speedup for the Rygg pinhole correction, because it was
taking a very long time to compute the tth distortion field for individual
detectors, but much less time to do so for the whole polar view grid.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>