Include position for rotation series angles
For some reason, the `valid_angs` coming out of `simulate_roation_series()`
are not modified when the grain position changes. I'm not sure if this is
expected behavior or not (see [here](https://github.com/HEXRD/hexrdgui/issues/1383#issuecomment-1501950293)).

This was causing the ranges for the overlays to not be updated correctly
in the raw and cartesian views when the grain positions were modified (although
the center spot was still updated correctly). And it also meant that both
the spots and ranges were not updated in the polar and stereo views.

Instead of using these angles, convert from the `valid_xys`, which do include
grain position info.

This means, however, that the omegas likely do not include any grain position
info as well (which may be an issue for non-aggregated views).

Possibly fixes #1383 (need confirmation after merging).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>