Do not apply template to completed TARDIS image
For the TARDIS workflow we need to load a new image for each detector. After
completing a detector the user will select the next detector and then load the
image. Since we now automatically apply the template on detector selection this
would result in the completed dashed outline being overlayed with another
template. To prevent confusion delay this in the case of TARDIS and do not
apply the template until the next image is loaded.

Signed-off-by: Brianna Major <brianna.major@kitware.com>