Fix places where `ArtistList.remove()` was used
`ArtistList.remove()` is no longer a method in matplotlib, as of 3.7.0
(it was made immutable, see [here](https://matplotlib.org/stable/api/axes_api.html#matplotlib.axes.Axes.ArtistList), and it may be turned into a tuple soon).

As such, we encounter an error in these places for new versions of matplotlib.

Instead, we can call `remove()` on the artist itself, and this removes it from
the `ArtistList` automatically. Do that instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>