Make pinhole mask dialog more interactive
The pinhole mask settings are now saved so that they persist both in the
same session and in between sessions.

Also, the dialog now stays open with an "Apply" button. When this is clicked,
the previous pinhole mask is replaced with a mask generated with the new
settings. This allows users to modify the values, create the pinhole mask, and
then repeat in case the values need to be modified.

Fixes: #1417

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>