Improve polar view mouse hover info
Before, x and y were being displayed in the main polar canvas, but these
were the same as `tth` and `eta`, which were also displayed, and was therefore
redundant.

Now, x and y are not displayed.

Also, in the lineout, x is now clearly labeled as `tth`, and the intensity as
well as `Q` are displayed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>