Do not update reflections table during selection
Reflections selection has been a bit buggy. It appears that the issue is that
updating the reflections table during selection messes up something with Qt's
selection model.

But we do not need to update the reflections table during selection. So do not
do it. This appears to fix the bugginess.

Fixes: #1367

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>