Do not reset material exclusions for export HEDM
As of hexrd/hexrd#513, the material exclusions are now reset by default
when running fit-grains.

In hexrd/hexrd#520, they are now reset in the materials as well. However,
that PR also introduces the option to not reset exclusions, both in fit-grains
and in the material config.

Set that option when we export the config so that we keep the exclusions
that we saved on the materials. This is so that we are able to reproduce what
was set in the GUI when running the HEDM workflow from the command line.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>