Add ability to enable/disable image mode widget
There are various times throughout the program where we need to disable
options in the image mode widget - especially during interactions with
the canvas such as in the LLNL import tool or during hand picking for
calibration.

Add the ability to enable/disable the image mode widget so that the user
cannot break the current interactions.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>