Ensure tree view gets fully reset when clearing
When we cleared the tree view before, we would call `beginRemoveRows()`
and `endRemoveRows()`. However, some parts of the model were apparently
not notified of the underlying data change, and they were not updating
properly. Maybe a `dataChanged()` would have fixed this, but since we
are deleting everything anyways, we should just do `beginResetModel()`
and `endResetModel()` instead.

This appears to fix the issues.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>