Fix line picker zoom box on canvas borders
When the line picker zoom box was previously near canvas borders,
the clipping to the x and y limits that would occur would make the
drawn zoom box image incorrect.

Instead of clipping the x and y limits, we should instead clip the
lines in the subplots. This fixes the issue.

Fixes: #1469

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>