Add a canvas focus mode for calibration
This is currently only used for structureless calibration, but it basically
disables several main window widgets, including the config toolbox (that
includes material and instrument config), the main menu bar, and the image
mode widget.

The widgets that are not disabled are the main canvas, the messages widget,
and the color map editor.

But we should use this focus mode very cautiously, and make sure it never
gets left on by accident, even if an exception occurs in the code, because
this will ruin the user's current session.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>