Fix polar mask when eta is split
If a polar mask is split in eta so that part of it is at the top and
part of it is at the bottom, it is now split into two polygons. These
polygons get some extra padding (so they cover all pixels to the eta
border) and some interpolation is done in the two theta pixel to
ensure the padding runs in the correct direction.

Fixes: #1369

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>