Always create polar masks from raw data
Instead of creating polar masks directly from the polar coordinates
at first, always create the polar masks from the raw data.

This means the polar coordinates get converted to raw coordinates,
and then back to polar coordinates, for generating the polar masks.

This is important for consistency. Users have been noticing
(such as in #1458), that if they switch views, sometimes the masks
vanish. This means the conversion to raw and back didn't work
properly. Not only that, but sometimes the pixels for the mask are
shifted a little bit after converting to raw and back.

So for consistency, always convert to raw and back. That way, the
users will not see any difference if they switch view types.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>