Flip the stereo view along the x-axis
The matplotlib origin by default is at the top left corner of the image.
Instead, put it at the bottom left corner of the image, so that the image
is flipped along the x-axis, and it appears more intuitive (as this is the
orientation that it is actually in during the experiment).

It appears that mouse hover info has been corrected as well with moving
the origin.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>