Generalize InteractiveTemplate class
Decouples the InteractiveTemplate class from the LLNL import tool so that it
can be used more broadly across the application. The class can now accept a
list of vertices in raw coordinates and create a polygon, rectangle, or
ellipse. If the static_mode attribute is set to False the user can translate or
rotate the shape.

Signed-off-by: Brianna Major <brianna.major@kitware.com>