Reorder async runner operations to handle errors
Since the async worker was running in a separate thread, if the task
encountered an error very fast, it would sometimes exit before we could
finish making connections. This would cause a segmentation fault.

We saw this in the structureless calibration when we would accidentally have
some variables where the min was greater than the max. The error would occur
quickly, before the connections were finished being made, and there would
be a segmentation fault.

Reorder so that the task doesn't start until the connections have finished
being made. This prevents the segmentation fault.

Also, accept the progress dialog later, because sometimes it would hang
around and never actually get closed. But it appears to close correctly now.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>