Do not rescale the cartesian view
We have been rescaling the cartesian view for quite a while, but
it does not actually appear to be necessary.

The pixel values from the raw view seem to be maintained when they
are transformed to the cartesian view. Rescaling the values in the
cartesian view is sometimes causing problems when there are, for
instance, very high and low values.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>