Allow going back a line while hand-picking points
Previously, if the user accidentally right-clicked, we would move on
to the next line, and there would be no way to go back to the previous
line. This would mess up their hand-picking.

Now, the "Back" button is enabled, even when the current line has no
points, and clicking "Back" when the current line has no points results
in going back a line instead.

Fixes: #1567

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>