Add an option to modify the font size of the app
This option is present in the new "Edit"->"Configuration"
dialog. This modifies the font size of both Qt and matplotlib.

When the user modifies the font size in the dialog, most (but not all)
widgets in the application are updated immediately. It appears that
any widgets with custom style sheets do not get updated. Perhaps some
other types of widgets don't get updated either. But restarting the
application normally appears to update everything to the new font
size.

It would be nice to get all widgets to update immediately when the
user modifies the font size. But for now, we can just present a warning
that the user should restart the application normally in order to get
all widgets to update.

Fixes: #1575

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>