Fix commitData warning in tree views
The following warning was appearing:

```
QAbstractItemView::commitData called with an editor that does not belong to this view
```

It looks like the modified functions were getting called with a checkbox
before it belonged to the view.

To fix it, we now only commit the data on user interaction (the `clicked` signal
instead of the `toggled` signal).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>