Fix invalid QSettings issue
It appears that PySide6 saves and loads QSettings a little differently
than PySide2. When we load QSettings into PySide6 that were generated
in PySide2, lists that are nested in dicts become lists of lists
instead of flat lists. There may be some other differences too.

Rather than trying to figure out how to patch the QSettings, let's
just make a new version of QSettings. This means that when users
upgrade to the PySide6 version, they will lose their settings, but
that is okay...

This doesn't affect state files at all, which do not use QSettings
but a different mechanism for storing state.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>