Hide QMessageBox instead of rejecting
We don't need to reject because `self.remove_lines()` is already
called. And rejecting is causing an issue due to a bug in Qt
(reported [here](https://bugreports.qt.io/browse/PYSIDE-2490)).

This issue is new to PySide6. We might run into it elsewhere. But
let's hope that Qt fixes the issue soon.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>