Use most recent images directory for image save
The save images dialog used to use the working directory as the default
location. This may not be in sync (like after loading a state file) and may not
be a good starting point. Instead use the most recent images directory. If that
path does not exist or is invalid default to the current working directory.

Signed-off-by: Brianna Major <brianna.major@kitware.com>