Do not set detailed text in message box
On Linux, the detailed text is hidden by default and can be shown via a
button.

On Mac, however, the detailed text is always shown, and if it is too
long, the buttons at the bottom of the dialog box will be unreachable
on the user's monitor.

Don't put the detailed text in the message box. We already print it
out in the messages widget, and that should be good enough. Only keep
the abbreviated message in the message box.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>