Add ability to copy reflections table to clipboard
If a user right-clicks a row in the reflections table, a context
menu appears with a single option: "Copy to Clipboard".

If the user clicks "Copy to Clipboard", the selected rows in the
reflections table will be copied to the clipboard in CSV format.
This may then be pasted anywhere (including to Excel).

An example:

```csv
ID,{hkl},d-spacing (Å),2θ (°),|F|²,Iₚ,Multiplicity
0,1 1 1,3.12,3.17,100.00,100.00,8
1,2 0 0,2.71,3.66,44.98,25.29,6
2,2 2 0,1.91,5.18,52.19,29.30,12
3,3 1 1,1.63,6.08,19.52,15.93,24
4,2 2 2,1.56,6.35,10.37,2.59,8
```

Fixes: #1617

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>