Add a cancel button to fast powder calibration
This generalizes the cancel button to work with the AsyncRunner class.
It also introduces a new CancelableProcess class to help make cancel
buttons simpler to implement in the future.

Threads are difficult to cancel. They can only really be canceled
appropriately if we have well-placed checks in places like loops to see
if the task was canceled. That can make the cancel-checking code quite
invasive. It also introduces delays in canceling, since the thread,
for instance, may need to finish an iteration of a loop before it checks again
if it was canceled.

Processes are easier to cancel, since we can instantly kill them. The operating
system can usually handle any clean-up for canceled processes quite well,
since it is common. We can also recursively kill all child processes, so it
works also if the task itself performs multiprocessing. So spawning new
processes is a much more flexible and less invasive way to cancel tasks.

However, it requires that the input and output be pickled and sent between
the main process and the child process. In some situations, the data may not
be pickle-able, and in some situations, there may be a large performance hit
caused by the pickling step. So using the CancelableProcess is only appropriate
if the inputs/outputs are pickleable, and they aren't large enough to cause
a noticeable performance hit.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>