Add option to lock ROI group detector transforms
If an ROI instrument config is loaded, a new checkbox appears in the
instrument slider view with the label "Lock ROI Group Transformations".

If checked, then when a detector is transformed (via translation or tilt),
all other detectors in the same group are transformed similarly.

For translation, all other detectors in the same group are translated by
the same amount, and in the same direction.

For tilt, all detectors in the group are rotated about the mean of their
centers by the tilt amount (which means their translation is affected too).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>