Fix issue where tilt parameters would jump around
During user interaction, the tilt parameters might jump around due to
non-uniqueness when converting to exponential map parameters and then back
to tilt parameters.

But we didn't need to convert to exponential map parameters for the original
detector to begin with anyways. We don't need to re-compute them based upon
the rotation matrix.

So just keep whatever value was set and don't recompute it. This fixes the
non-uniqueness jumpiness.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>