Keep two copies of the polar data
One copy is used internally for computations and applies all masks - "visible"
and "boundary". The other is the display image and is used for the generated
images (polar, stereo, azimuthal lineout). The display image only applies the
"visible" masks and "boundary" masks are only drawn on top of the image.

Signed-off-by: Brianna Major <brianna.major@kitware.com>