Set panel buffer values to 0 before polar view
For Eiger, the invalid pixels have a value of 4294967295. This high
value bleeds into neighboring pixels when we warp to the polar view
and makes them artificially high.

To prevent this bleeding over, we should set panel buffer values to 0
before generating the polar view.

0 seemed to be better than nan because nan would result in more polar view
pixels being nan. Maybe an even better setup, however, would be to make
the invalid pixels be the average value of their neighbors, so that they
do not bring down the neighboring values in the final image. Something to
think about in the future...

Note that masks in hexrdgui are not set to zero before generating the
polar view, so the masked content may still bleed into neighboring pixels.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>