Avoid creating extra computed images if possible
This adds a property to the mask manager that we can use to determine
whether or not there are border-only masks present.

If there are border-only masks present, then the display images and
computed images will be different, so we will have to re-create the
computed images from scratch.

If there are no border-only masks, then the display images and the
computed images are the same, and we can skip that extra computation
step by re-using the display images as the computed images.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>