Aggregate HKL powder lines in the polar lineout
Before, each detector would draw the mean of its own powder line data
in the polar view azimuthal lineout. This wasn't noticeable when the
powder lines are completely straight. However, when the powder lines
are not completely straight, the duplicate lines would be visible as
separate means.

To avoid this, concatenate all lines that belong to the same HKL from
each detector, and take their mean together. Use this mean to draw a
single line in the azimuthal lineout. Repeat this kind of thing for
the ranges, too.

Fixes: #1432

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>