Improve performance of generating polar masks
For a large number of polar masks, `create_hedm_instrument()` was
taking a lot of cumulative time, since it was being called repeatedly.
Provide the option to pass it in as an argument for performance.

Also, an intensity-corrected images dict was being repeatedly created.
Now, we skip that step if the PolarView is initialized with an instrument
that is `None`.

This resulted in a ~50% performance improvement for generating the polar
view with an example that had 42 masks.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>