Convert raw masks to polar with contouring
This creates a polygon and performs contouring to convert raw masks
to polar. This fixes a number of issues we have seen by our attempted
enhancements at the previous methods.

This also switches over to using Pillow's method for drawing polygons,
since it is substantially faster than scikit image's method (despite
requiring a list for input).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>