Do not rerender when overwriting loaded state file
We were previously rerendering, because overwriting a loaded state
file requires us to reload the imageseries from the new state file, and
we wanted to make sure that any changes would be reflected in the UI.

However, there really shouldn't be any changes, and rerendering sometimes
takes quite a while (for example, in high resolution polar view). So just
skip rerendering, and do not allow the colormap ranges to reset either.

Fixes: #1694

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>