Add option to set lmfit boundaries as a delta
lmfit does not allow a delta to be used, but instead a min/max must
be specified. We can, however, allow a delta to be used in the GUI,
and then set the min/max on the lmfit parameters based upon this delta.
That is what is being done in this PR.

Using delta for boundaries is more intuitive for some workflows.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>