Filter down engineering constraint types
We are currently identifying TARDIS and PXRDIP both when
`guess_instrument_type()` is used. However, only TARDIS has
engineering constraints, and an exception is raised if we try
to set engineering constraints using PXRDIP.

When guessing the engineering constraints, we should filter down
(ignore) any instrument types for which we do not yet have
engineering constraints defined. This prevents errors from occuring
later.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>