Do not rescale from WPPF scatter points
Previously, if the user ran WPPF without making "Scale" refinable
in the first iteration, the azimuthal integration scale would become
very zoomed out (in y) since the WPPF points would often be extremely
large. This would be difficult to fix since future refinements did
not scale the y-axis down.

This commit adds an autoscaling that ignores the WPPF scatter points, and
that keeps it zoomed in on the line data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>