Add a point picker to use for spline in WPPF
The spline method requires picking background points. This was previously
being done with pylab in hexrd. However, this is a deprecated module, and
it required the use of middle click (which not everyone had access to). It
also would make points appear while doing things such as pan/zoom.

The new point picker solves all of these issues.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>