Add initial support for FIDDLE instrument
Adds a new load path option to the LLNL Import dialog to support the basic needs
of the FIDDLE instrument. This skips the template/masking step altogether and
instead allows a user to select an image and dark file for a detector. The
remaining detector image and dark files are automatically found and associated
with the correct detector.

Signed-off-by: Brianna Major <brianna.major@taloid.khq.kitware.com>