HEDM Calibration: restore grain settings
In the HEDM calibration, if the user does things like unchecking whether
the strain is fixed to identity, restore the previous setting that the
strain had.

This way, modifying options won't result in a permanent loss of the
previous settings. It also enables us to fix the strain by default
(which was recommended).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>