Add options for locking slider transforms
Previously, slider transforms could only be locked if ROI subpanels
with groups were used.

Now, slider transforms can be locked regardless. An
"Instrument Rigid Body" lock will results in the entire instrument
being transformed together. "Group Rigid Body" lock is only available
if every detector has a specified group, and it allows detectors within
a group to be transformed together.

Additionally, the center of rotation may be specified. It defaults to
the "Mean Center", which is the mean center of the instrument (for
"Instrument Rigid Body") or the mean center of the group (for "Group
Rigid Body"). "Origin" is the lab origin at (0, 0, 0).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>