Avoid nested functions in QueuedConnections
The one in the indexing runner in particular was freezing the application
starting in Qt 6.8.1. I'm not entirely sure why, but making the function
a method on the class rather than making it a nested function appears to
fix the issue.

We'll just try to avoid using nested functions in Qt QueuedConnections.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>