Fix matplotlib version to be less than 3.10
There's some new issue with removing several of our artists starting
in matplotlib 3.10 (including overlays, detector borders, the beam
marker, and potentially more). It is very common that when `remove()`
is called on the artists, that a `NotImplementedError` is raised.

Perhaps this is an issue that will get fixed in matplotlib soon.
If not, we'll need to figure out an alternative way to remove the
artists.

But for the release coming up, let's just fix the matplotlib version.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>