Do not update detector coatings for removed detectors
This is for backwards compatability. Previously, when detector coatings were
set after using the LLNL import tool the "default" detector (used to represent
the current detector during import) was not removed after import was complete.
This means that some old state files will have this value and try to fetch this
detector's information, even though it no longer exists.

Signed-off-by: Brianna Major <brianna.major@taloid.khq.kitware.com>