Disable non-selected x-ray source pick paths
If "Show picks from all x-ray sources" is not checked, picks for
the non-selected x-ray source are now grayed out and not editable.
This is much better than before, where an error would simply occur
if users tried to edit non-selected x-ray source picks (or worse,
in structureless calibration, the point would be added inadvertently
to the wrong source).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>