Fix event check for translate/rotate template
We use NaNs in our template vertices in order to draw "holes" within an outer
boundary. When we check a click event to determine if a user has selected the
template or not we should only check if the outer boundary contains the event.
Using the full polygon causes some weird and inconsistent behaviors.

Signed-off-by: Brianna Major <brianna.major@kitware.com>