Refactor exporting the polar view
We were previously exporting raw intensities as `intensities`. Now,
we export the exact intensities displayed in the polar view as
`intensities`, and the raw intensities are exported as `raw_intensities`.

This also adds several additional arrays that get exported from the
polar view, including the erosion mask, tth correction field, and
border masks. Overall, this adds a lot of useful arrays to the polar
view export.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>